backend/backend/ragapp/rag-app/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
requests==2.26.0
beautifulsoup4==4.10.0
chroma==0.1.0
openai
langchain==0.0.1
bs4==0.0.1
python-dotenv>=1.0.1,<2.0.0
python-decouple==3.8
pymilvus==2.4.12
langchain-openai
PyMuPDF==1.22.3


